requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
qdrant-client
langchain
httpx
supabase
pymupdf
pydantic-settings
python-multipart
langchain-community
tiktoken
langchain_openai
python-jose[cryptography]
slowapi
pydantic[email]